Create TextBox Component and install font-awesome
The textbox component can used for login and register form page. It is
designed for a text based input tag (not radio, checkbox, or textarea).
You must define the textbox props with an object consist of name, type
placeholder, and icon(fa-meh-o).

You will also got a font-awesome installed at this commit.